✨ Link template and devices
By default, templates are not linked to the device they depend on (if
any). This has been requested several times (see
https://community.home-assistant.io/t/add-template-sensors-to-device-page/193413).

Main use case is to associate a virtual sensor defined using a template
to the device it refers to.

This patch adds a device for each template defined using HA UI and links
this device to the original device if relevant.
It works only for templates that depend only on entities from a single
device.

Implementation detail:

* `device_info` is called only once, just when adding the entity to HA
* it happens before template is executed for the first time, and thus
  before we have a list of entites a template depends on (which make
  change during the lifetime of the template by the way)
* it forces us to use the "intermediate" device trick and to link this
  device to the real one when we have executed the template for the
  first time.

Reference: https://community.home-assistant.io/t/add-template-sensors-to-device-page/193413/8